+++ src/collective/sortedlisting/browser/static/sortablequerystring/pattern.js
@@ -41,8 +41,9 @@
 define([
   'jquery',
   'mockup-patterns-querystring',
-  'mockup-patterns-sortable'
-], function($, QueryString, Sortable) {
+  'mockup-patterns-sortable',
+  'translate'
+], function($, QueryString, Sortable, _t) {
   'use strict';
 
   var SortableQueryString = QueryString.extend({
@@ -90,8 +91,20 @@ define([
         query.push('sort_order=reverse');
       }
 
-      var sorting = $("textarea[name$='.sorting']" ).val();
-      if (sorting !== undefined) {
+      /* Get active modal window. Due to the properties-modal (which is always in the DOM but hidden if not active) we
+       * have 2 modal windows in the DOM when a tile-modal is active. This leads to inconsistencies when the
+       * SortableCollectionBehavior & the Sortable-Content-listing-Tile are active. */
+      var modal = $('div.plone-modal-wrapper:visible').first();
+      if(modal.length == 0) {
+        modal = $('div.plone-modal-wrapper.mosaic-overlay');
+      }
+      if(modal.length == 0){
+        modal = $('div#content-core');
+      }
+
+      var sorting = modal.find("textarea[name$='.sorting']").first().val();
+
+      if (sorting !== undefined && self.$sortOn.val() === '') {
         query.push('sorting=' + sorting.split('\n').join(','));
       }
 
@@ -102,18 +115,88 @@ define([
               .html(data)
               .appendTo(self.$previewPane);
 
-            var uidList = $("#search-results li").map(function() {
+            var uidList = modal.find("div#search-results li").map(function() {
               return $(this).data("uid");
             }).get();
-            $( "textarea[name$='.sorting']" ).val( uidList.join("\r\n"))
 
+            modal.find("textarea[name$='.sorting']").first().val( uidList.join("\r\n"));
 
-            var dd = new Sortable($('#search-results'), {
+            var dd = new Sortable(modal.find('div#search-results ul.sortedListing-results').first(), {
                selector: 'li',
                drop: 'updateSorting'
             });
-
           });
+    },
+    createSort: function() {
+      var self = this;
+
+      var modal = $('div.plone-modal-wrapper:visible').first();
+      if(modal.length == 0) {
+        modal = $('div.plone-modal-wrapper.mosaic-overlay');
+      }
+      if(modal.length == 0){
+        modal = $('div#content-core');
+      }
+
+      // elements that may exist already on the page
+      // XXX do this in a way so it'll work with other forms will work
+      // as long as they provide sort_on and sort_reversed fields in z3c form
+      var existingSortOn = modal.find('[id$="-sort_on"]').filter('[id^="formfield-"]');
+      var existingSortOrder = modal.find('[id$="-sort_reversed"]').filter('[id^="formfield-"]');
+
+      $('<span/>')
+        .addClass(self.options.classSortLabelName)
+        .html(_t('Sort on'))
+        .appendTo(self.$sortWrapper);
+      self.$sortOn = $('<select/>')
+        .attr('name', 'sort_on')
+        .appendTo(self.$sortWrapper)
+        .change(function() {
+          self.refreshPreviewEvent.call(self);
+          $('[id$="sort_on"]', existingSortOn).val($(this).val());
+        });
+
+      self.$sortOn.append($('<option value="">No sorting</option>')); // default no sorting
+      for (var key in self.options['sortable_indexes']) { // jshint ignore:line
+        self.$sortOn.append(
+          $('<option/>')
+            .attr('value', key)
+            .html(self.options.indexes[key].title)
+        );
+      }
+      self.$sortOn.patternSelect2({width: '150px'});
+
+      self.$sortOrder = $('<input type="checkbox" />')
+        .attr('name', 'sort_reversed:boolean')
+        .change(function() {
+          self.refreshPreviewEvent.call(self);
+          if ($(this).prop('checked')) {
+            $('.option input[type="checkbox"]', existingSortOrder).prop('checked', true);
+          } else {
+            $('.option input[type="checkbox"]', existingSortOrder).prop('checked', false);
+          }
+        });
+
+      $('<span/>')
+        .addClass(self.options.classSortReverseName)
+        .appendTo(self.$sortWrapper)
+        .append(self.$sortOrder)
+        .append(
+          $('<span/>')
+            .html(_t('Reversed Order'))
+            .addClass(self.options.classSortReverseLabelName)
+        );
+
+      if (existingSortOn.length >= 1 && existingSortOrder.length >= 1) {
+        var reversed = $('.option input[type="checkbox"]', existingSortOrder).prop('checked');
+        var sortOn = $('[id$="-sort_on"]', existingSortOn).val();
+        if (reversed) {
+          self.$sortOrder.prop('checked', true);
+        }
+        self.$sortOn.select2('val', sortOn);
+        $(existingSortOn).hide();
+        $(existingSortOrder).hide();
+      }
     },
   });