+++ CHANGES.rst
@@ -5,7 +5,16 @@ Changelog
 1.0a5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Preview-field is now bigger (drag & drop usability).
+- Same results-limit will be used in the view and the querybuilder.
+- The sorting for every collection will be stored in it's own sorting field.
+- "Sort on"-field has now an impact on the query & the results.
+- "Sort on"- and "Sort order"-fields will be stored separate for every collection.
+- The sorting will be updated always. Before this happend only while saving.
+- Set automatically "Sort on"-field to None (which is necessary for d&d) when d&d is executed.
+- Show correct sorting also if limit is smaller than the amount of results.
+- Fix tests
+  [adrianschulz]
 
 
 1.0a4 (2017-11-09)